requirements.txt
# Django
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0

# Photo & Image
Pillow==9.1.0